[unrated] Title: 배열 만들기 5, Time: 3.22 ms, Memory: 96.1 MB -BaekjoonHub

+++ 프로그래머스/unrated/181912. 배열 만들기 5/배열 만들기 5.java
@@ -15,7 +15,7 @@ class Solution {
     private int[] toArray(ArrayList<Integer> list) {
         int[] result = new int[list.size()];
         for (int i = 0; i < result.length; i++)
-            result[i] = list.get(i).intValue();
+            result[i] = list.get(i);
         
         return result;
     }
